feat: created method for add personal spends refactor: changed the way data is saved

+++ src/main/java/ru/daniil/split/model/Person.java
@@ -1,22 +1,23 @@
 package ru.daniil.split.model;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
 public class Person {
 
     public final String name;
-    private final List<Integer> allSpends = new ArrayList<>();
+    private final List<BigDecimal> allSpends = new ArrayList<>();
 
     public Person(String name) {
         this.name = name;
     }
 
-    public List<Integer> getAllSpends() {
+    public List<BigDecimal> getAllSpends() {
         return new ArrayList<>(allSpends);
     }
 
-    public void addSpend(Integer newSpend) {
+    public void addSpend(BigDecimal newSpend) {
         this.allSpends.add(newSpend);
     }
 }
